leila_online_agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.40.0
pydantic==2.8.2
SQLAlchemy==2.0.31
python-dotenv==1.0.1
